Flytt alt av papirting til papirpakke
